API/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.1
pyngrok>=5.1.0
python-docx>=0.8.11
openpyxl>=3.0.9
PyPDF2>=2.6.0
tika>=1.24
pandas>=1.3.3
pytest>=6.2.5
httpx==0.13.3
python-multipart>=0.0.5
aiofiles>=0.7.0 